requirements.txt
flask==2.0.1
flask-cors==3.0.10
nltk==3.6.3
python-dotenv==0.19.0
pymongo==3.12.0
requests==2.26.0